pyproject.toml
# ******************** project metadata ********************#
[project]
name = "teaching-assistant"
version = "0.1.0"
description = "use ai be fast"
readme = "README.md"
requires-python = ">=3.13"

# ******************** core dependencies ********************#
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "httpx",
    "python-dotenv",
    "crawl4ai",
    "playwright",
    "asyncio",
    "html2text>=2025.4.15",
    "tiktoken>=0.9.0",
    "typesense>=1.1.1",
    "openai>=1.82.1",
    "tesseract>=0.1.3",
    "pytesseract>=0.3.13",
    "pyspellchecker>=0.8.3",
    "pyyaml>=6.0.2",
    "ollama>=0.5.1",
    "beautifulsoup4>=4.13.4",
    "pillow>=10.4.0",
    "requests>=2.32.3",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "easyocr>=1.7.2",
    "opencv-python-headless>=4.8.0",
    "typing-extensions>=4.13.2",
    "langchain-text-splitters>=0.3.8",
    "sentence-transformers>=4.1.0",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "tqdm>=4.67.1",
    "pytest>=8.4.0",
]

# ******************** torch configuration ********************#
[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
]

# ******************** pytorch index configuration ********************#
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
]
